registration and authorization is worked

+++ TourAgency/TourAgency/Entities/Taxation.cs
@@ -18,7 +18,7 @@ namespace TourAgency.Entities
         [DatabaseGenerated(DatabaseGeneratedOption.None)]
         public int Id { get; set; }
 
-        public decimal? Tax { get; set; }
+        public double? Tax { get; set; }
 
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<Client> Client { get; set; }
